Re #20240

+++ scripts/test/SANS/gui_logic/run_tab_presenter_test.py
@@ -1,22 +1,52 @@
+
 from __future__ import (absolute_import, division, print_function)
 
 import unittest
 import sys
 
-import mantid
 from mantid.kernel import config
 from mantid.kernel import PropertyManagerDataService
 
 from sans.gui_logic.presenter.run_tab_presenter import RunTabPresenter
-from sans.common.enums import (SANSFacility, ReductionDimensionality, SaveType, OutputMode, ISISReductionMode,
+from sans.common.enums import (SANSFacility, ReductionDimensionality, SaveType, ISISReductionMode,
                                RangeStepType, FitType)
-from sans.test_helper.user_file_test_helper import (create_user_file, sample_user_file)
+from sans.test_helper.user_file_test_helper import (create_user_file, sample_user_file, sample_user_file_gravity_OFF)
 from sans.test_helper.mock_objects import (create_mock_view)
 from sans.test_helper.common import (remove_file, save_to_csv)
+
+
 if sys.version_info.major == 3:
     from unittest import mock
 else:
     import mock
+
+
+BATCH_FILE_TEST_CONTENT_1 = "# MANTID_BATCH_FILE add more text here\n" \
+                            "sample_sans,1,sample_trans,2,sample_direct_beam,3," \
+                            "output_as,test_file,user_file,user_test_file\n" \
+                            "sample_sans,1,can_sans,2,output_as,test_file2\n"
+
+
+BATCH_FILE_TEST_CONTENT_2 = "# MANTID_BATCH_FILE add more text here\n" \
+                            "sample_sans,SANS2D00022024,sample_trans,SANS2D00022048," \
+                            "sample_direct_beam,SANS2D00022048,output_as,test_file\n" \
+                            "sample_sans,SANS2D00022024,output_as,test_file2\n"
+
+BATCH_FILE_TEST_CONTENT_3 = "# MANTID_BATCH_FILE add more text here\n" \
+                            "sample_sans,1p3,output_as,test_file\n"
+
+
+class MultiPeriodMock(object):
+    def __init__(self, call_pattern):
+        self._counter = 0
+        self._call_pattern = call_pattern
+
+    def __call__(self):
+        if self._counter < len(self._call_pattern):
+            return_value = self._call_pattern[self._counter]
+            self._counter += 1
+            return return_value
+        raise RuntimeError("Issue with multi-period mocking.")
 
 
 class RunTabPresenterTest(unittest.TestCase):
@@ -61,7 +91,7 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertTrue(view.transmission_mask_files == "test4.xml")
         self.assertTrue(view.transmission_radius == 7.)
         self.assertTrue(view.transmission_monitor == 4)
-        self.assertTrue(view.transmission_m4_shift == -70)
+        self.assertTrue(view.transmission_mn_shift == -70)
         self.assertTrue(view.transmission_sample_use_fit)
         self.assertTrue(view.transmission_sample_fit_type is FitType.Logarithmic)
         self.assertTrue(view.transmission_sample_polynomial_order == 2)
@@ -72,9 +102,7 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertFalse(view.pixel_adjustment_det_2)
         self.assertFalse(view.wavelength_adjustment_det_1)
         self.assertFalse(view.wavelength_adjustment_det_2)
-        self.assertTrue(view.q_1d_min == 0.001)
-        self.assertTrue(view.q_1d_max == 0.2)
-        self.assertTrue(view.q_1d_step == 0.001)
+        self.assertTrue(view.q_1d_min_or_rebin_string == "0.001,0.001,0.0126,-0.08,0.2")
         self.assertTrue(view.q_xy_max == 0.05)
         self.assertTrue(view.q_xy_step == 0.001)
         self.assertTrue(view.q_xy_step_type == RangeStepType.Lin)
@@ -89,11 +117,20 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertTrue(view.radius_limit_min == 12.)
         self.assertTrue(view.radius_limit_min == 12.)
         self.assertTrue(view.radius_limit_max == 15.)
+        self.assertFalse(view.compatibility_mode)
+        self.assertFalse(view.show_transmission)
+
+        # Assert that Beam Centre View is updated correctly
+        self.assertEqual(view.beam_centre.lab_pos_1, 155.45)
+        self.assertEqual(view.beam_centre.lab_pos_2, -169.6)
+        self.assertEqual(view.beam_centre.hab_pos_1, 155.45)
+        self.assertEqual(view.beam_centre.hab_pos_2, -169.6)
 
         # Assert certain function calls
         self.assertTrue(view.get_user_file_path.call_count == 3)
         self.assertTrue(view.get_batch_file_path.call_count == 2)  # called twice for the sub presenter updates (masking table and settings diagnostic tab)  # noqa
-        self.assertTrue(view.get_cell.call_count == 36)
+        self.assertTrue(view.get_cell.call_count == 64)
+
         self.assertTrue(view.get_number_of_rows.call_count == 6)
 
         # clean up
@@ -112,34 +149,63 @@ class RunTabPresenterTest(unittest.TestCase):
             has_raised = True
         self.assertFalse(has_raised)
 
-    def test_that_loads_batch_file_and_places_it_into_table(self):
+    def do_test_that_loads_batch_file_and_places_it_into_table(self, use_multi_period):
         # Arrange
-        content = "# MANTID_BATCH_FILE add more text here\n" \
-                  "sample_sans,1,sample_trans,2,sample_direct_beam,3," \
-                  "output_as,test_file,user_file,user_test_file\n" \
-                  "sample_sans,1,can_sans,2,output_as,test_file2\n"
-        batch_file_path = save_to_csv(content)
-        user_file_path = create_user_file(sample_user_file)
-        view, settings_diagnostic_tab, masking_table = create_mock_view(user_file_path, batch_file_path)
-        presenter = RunTabPresenter(SANSFacility.ISIS)
-        presenter.set_view(view)
+        batch_file_path, user_file_path, presenter, view = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_1,
+                                                                                              is_multi_period=use_multi_period)  # noqa
 
         # Act
         presenter.on_batch_file_load()
 
         # Assert
         self.assertTrue(view.add_row.call_count == 2)
-        expected_first_row = "SampleScatter:1,SampleTransmission:2,SampleDirect:3," \
-                             "CanScatter:,CanTransmission:,CanDirect:,OutputName:test_file"
-        expected_second_row = "SampleScatter:1,SampleTransmission:,SampleDirect:," \
-                              "CanScatter:2,CanTransmission:,CanDirect:,OutputName:test_file2"
+        if use_multi_period:
+            expected_first_row = "SampleScatter:1,ssp:,SampleTrans:2,stp:,SampleDirect:3,sdp:," \
+                                 "CanScatter:,csp:,CanTrans:,ctp:,CanDirect:,cdp:,OutputName:test_file"
+            expected_second_row = "SampleScatter:1,ssp:,SampleTrans:,stp:,SampleDirect:,sdp:," \
+                                  "CanScatter:2,csp:,CanTrans:,ctp:,CanDirect:,cdp:,OutputName:test_file2"
+        else:
+            expected_first_row = "SampleScatter:1,SampleTrans:2,SampleDirect:3," \
+                                 "CanScatter:,CanTrans:,CanDirect:,OutputName:test_file"
+            expected_second_row = "SampleScatter:1,SampleTrans:,SampleDirect:," \
+                                  "CanScatter:2,CanTrans:,CanDirect:,OutputName:test_file2"
 
         calls = [mock.call(expected_first_row), mock.call(expected_second_row)]
         view.add_row.assert_has_calls(calls)
 
         # Clean up
-        remove_file(batch_file_path)
-        remove_file(user_file_path)
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+    def test_that_loads_batch_file_and_places_it_into_table(self):
+        self.do_test_that_loads_batch_file_and_places_it_into_table(use_multi_period=True)
+
+    def test_that_loads_batch_file_and_places_it_into_table_when_not_multi_period_enabled(self):
+        self.do_test_that_loads_batch_file_and_places_it_into_table(use_multi_period=False)
+
+    def test_that_loads_batch_file_with_multi_period_settings(self):
+        # Arrange
+        batch_file_path, user_file_path, presenter, view = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_3,
+                                                                                              is_multi_period=False)
+        # The call pattern is
+        # False -- we are in single mode
+        # True -- we switch to multi-period mode
+        # True -- Getting table model
+        # True -- Getting table model
+        multi_period_mock = MultiPeriodMock(call_pattern=[False, True, True, True])
+        view.is_multi_period_view = mock.MagicMock(side_effect=multi_period_mock)
+
+        # Act
+        presenter.on_batch_file_load()
+
+        # Assert
+        self.assertTrue(view.add_row.call_count == 1)
+        self.assertTrue(view.set_multi_period_view_mode.call_count == 1)
+
+        expected_row = "SampleScatter:1,ssp:3,SampleTrans:,stp:,SampleDirect:,sdp:," \
+                       "CanScatter:,csp:,CanTrans:,ctp:,CanDirect:,cdp:,OutputName:test_file"
+
+        calls = [mock.call(expected_row)]
+        view.add_row.assert_has_calls(calls)
 
     def test_fails_silently_when_batch_file_does_not_exist(self):
         presenter = RunTabPresenter(SANSFacility.ISIS)
@@ -155,21 +221,11 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertFalse(has_raised)
 
         # Clean up
-        remove_file(user_file_path)
+        self._remove_files(user_file_path=user_file_path)
 
     def test_that_gets_states_from_view(self):
         # Arrange
-        content = "# MANTID_BATCH_FILE add more text here\n" \
-                  "sample_sans,SANS2D00022024,sample_trans,SANS2D00022048," \
-                  "sample_direct_beam,SANS2D00022048,output_as,test_file\n" \
-                  "sample_sans,SANS2D00022024,output_as,test_file2\n"
-        batch_file_path = save_to_csv(content)
-        user_file_path = create_user_file(sample_user_file)
-        view, _, _ = create_mock_view(user_file_path, batch_file_path)
-
-        presenter = RunTabPresenter(SANSFacility.ISIS)
-        presenter.set_view(view)
-
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
         presenter.on_user_file_load()
         presenter.on_batch_file_load()
 
@@ -207,24 +263,16 @@ class RunTabPresenterTest(unittest.TestCase):
         # Check some entries
         self.assertTrue(state0.slice.start_time is None)
         self.assertTrue(state0.slice.end_time is None)
+
         self.assertTrue(state0.reduction.reduction_dimensionality is ReductionDimensionality.OneDim)
+        self.assertEqual(state0.move.detectors['LAB'].sample_centre_pos1, 0.15544999999999998)
 
         # Clean up
-        remove_file(batch_file_path)
-        remove_file(user_file_path)
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
 
     def test_that_can_get_state_for_index_if_index_exists(self):
         # Arrange
-        content = "# MANTID_BATCH_FILE add more text here\n" \
-                  "sample_sans,SANS2D00022024,sample_trans,SANS2D00022048," \
-                  "sample_direct_beam,SANS2D00022048,output_as,test_file\n" \
-                  "sample_sans,SANS2D00022024,output_as,test_file2\n"
-        batch_file_path = save_to_csv(content)
-        user_file_path = create_user_file(sample_user_file)
-        view, _, _ = create_mock_view(user_file_path, batch_file_path)
-
-        presenter = RunTabPresenter(SANSFacility.ISIS)
-        presenter.set_view(view)
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
 
         presenter.on_user_file_load()
         presenter.on_batch_file_load()
@@ -241,23 +289,32 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertTrue(state.data.can_direct is None)
 
         # Clean up
-        remove_file(batch_file_path)
-        remove_file(user_file_path)
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+    def test_that_can_get_states_from_row_user_file(self):
+        # Arrange
+        row_user_file_path = create_user_file(sample_user_file_gravity_OFF)
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2,
+                                                                                           row_user_file_path = row_user_file_path)
+
+        presenter.on_user_file_load()
+        presenter.on_batch_file_load()
+
+        # Act
+        state = presenter.get_state_for_row(1)
+        state0 = presenter.get_state_for_row(0)
+
+        # Assert
+        self.assertTrue(state.convert_to_q.use_gravity is False)
+        self.assertTrue(state0.convert_to_q.use_gravity is True)
 
     def test_that_returns_none_when_index_does_not_exist(self):
         # Arrange
-        # Arrange
-        content = "# MANTID_BATCH_FILE add more text here\n" \
-                  "sample_sans,SANS2D00022024,sample_trans,SANS2D00022048," \
-                  "sample_direct_beam,SANS2D00022048,output_as,test_file\n" \
-                  "sample_sans,SANS2D00022024,output_as,test_file2\n"
-        batch_file_path = save_to_csv(content)
+        batch_file_path = save_to_csv(BATCH_FILE_TEST_CONTENT_2)
         user_file_path = create_user_file(sample_user_file)
         view, _, _ = create_mock_view(user_file_path, batch_file_path)
-
         presenter = RunTabPresenter(SANSFacility.ISIS)
         presenter.set_view(view)
-
         presenter.on_user_file_load()
         presenter.on_batch_file_load()
 
@@ -274,16 +331,7 @@ class RunTabPresenterTest(unittest.TestCase):
     def test_that_populates_the_property_manager_data_service_when_processing_is_called(self):
         # Arrange
         self._clear_property_manager_data_service()
-        content = "# MANTID_BATCH_FILE add more text here\n" \
-                  "sample_sans,SANS2D00022024,sample_trans,SANS2D00022048," \
-                  "sample_direct_beam,SANS2D00022048,output_as,test_file\n" \
-                  "sample_sans,SANS2D00022024,output_as,test_file2\n"
-        batch_file_path = save_to_csv(content)
-        user_file_path = create_user_file(sample_user_file)
-        view, _, _ = create_mock_view(user_file_path, batch_file_path)
-
-        presenter = RunTabPresenter(SANSFacility.ISIS)
-        presenter.set_view(view)
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
 
         # This is not the nicest of tests, but better to test this functionality than not
         presenter.on_user_file_load()
@@ -297,16 +345,120 @@ class RunTabPresenterTest(unittest.TestCase):
         self.assertTrue(len(PropertyManagerDataService.getObjectNames()) == 2)
 
         # clean up
-        remove_file(sample_user_file)
-        remove_file(user_file_path)
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
 
         self._clear_property_manager_data_service()
+
+    def test_that_calls_halt_process_flag_if_user_file_has_not_been_loaded_and_process_is_run(self):
+        # Arrange
+        self._clear_property_manager_data_service()
+        batch_file_path, user_file_path, presenter, view = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
+
+        presenter.on_processed_clicked()
+
+        # Assert
+        # We should have printed an error message to the logs and called halt process flag
+        self.assertTrue(view.halt_process_flag.call_count == 1)
+        # clean up
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+        self._clear_property_manager_data_service()
+
+    def test_that_calls_halt_process_flag_if_state_are_invalid_and_process_is_run(self):
+        # Arrange
+        self._clear_property_manager_data_service()
+        batch_file_path, user_file_path, presenter, view = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
+
+        presenter.on_batch_file_load()
+        presenter.on_user_file_load()
+
+        # Set invalid state
+        presenter._state_model.event_slices = 'Hello'
+
+        presenter.on_processed_clicked()
+
+        # Assert
+        # We should have printed an error to logs and called halt process flag
+        self.assertTrue(view.halt_process_flag.call_count == 1)
+
+        # clean up
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+        self._clear_property_manager_data_service()
+
+    def test_that_calls_halt_process_flag_if_states_are_not_retrievable_and_process_is_run(self):
+        # Arrange
+        self._clear_property_manager_data_service()
+        batch_file_path, user_file_path, presenter, view = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
+
+        presenter.on_batch_file_load()
+        presenter.on_user_file_load()
+
+        presenter.get_states = mock.MagicMock(return_value='')
+
+        presenter.on_processed_clicked()
+
+        # Assert
+        # We should have printed an error to logs and called halt process flag
+        self.assertTrue(view.halt_process_flag.call_count == 1)
+
+        # clean up
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+        self._clear_property_manager_data_service()
+
+    def test_that_can_add_new_masks(self):
+        # Arrange
+        self._clear_property_manager_data_service()
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_2)
+
+        presenter.on_user_file_load()
+        presenter.on_batch_file_load()
+
+        # Act
+        presenter.on_mask_file_add()
+
+        # Assert
+        state = presenter.get_state_for_row(0)
+        mask_info = state.mask
+        mask_files = mask_info.mask_files
+        self.assertTrue(mask_files == [user_file_path])
+
+        # clean up
+        self._remove_files(user_file_path=user_file_path, batch_file_path=batch_file_path)
+
+    def test_that_get_processing_options_returns_correct_value(self):
+        batch_file_path, user_file_path, presenter, _ = self._get_files_and_mock_presenter(BATCH_FILE_TEST_CONTENT_1)
+        expected_result = {'UseOptimizations':'1','OutputMode':'PublishToADS','PlotResults':'1','OutputGraph':'SANS-Latest'}
+
+        result = presenter.get_processing_options()
+
+        self.assertEqual(expected_result, result)
 
     @staticmethod
     def _clear_property_manager_data_service():
         for element in PropertyManagerDataService.getObjectNames():
             if PropertyManagerDataService.doesExist(element):
                 PropertyManagerDataService.remove(element)
+
+    @staticmethod
+    def _get_files_and_mock_presenter(content, is_multi_period=True, row_user_file_path = ""):
+        batch_file_path = save_to_csv(content)
+        user_file_path = create_user_file(sample_user_file)
+        view, _, _ = create_mock_view(user_file_path, batch_file_path, row_user_file_path)
+        # We just use the sample_user_file since it exists.
+        view.get_mask_file = mock.MagicMock(return_value=user_file_path)
+        view.is_multi_period_view = mock.MagicMock(return_value=is_multi_period)
+        presenter = RunTabPresenter(SANSFacility.ISIS)
+        presenter.set_view(view)
+        return batch_file_path, user_file_path, presenter, view
+
+    @staticmethod
+    def _remove_files(user_file_path=None, batch_file_path=None):
+        if user_file_path:
+            remove_file(user_file_path)
+        if batch_file_path:
+            remove_file(batch_file_path)
 
 
 if __name__ == '__main__':
